Added Following API

+++ src/main/java/com/vv/resource/TweetResource.java
@@ -11,6 +11,8 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
+import org.apache.commons.logging.Log;
+
 import com.vv.dao.TweetsDAO;
 import com.vv.dao.TweetsDAOImpl;
 import com.vv.model.Tweet;
@@ -27,7 +29,7 @@ public class TweetResource {
 	@GET
 	@Path("/{tweetId}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public Tweet getTweet(@PathParam("tweetId") long tweetId) {
+	public Tweet getTweet(@PathParam("tweetId") int tweetId) {
 		Tweet tweet = new Tweet();
 		tweet.setTweetId(1);
 		tweet.setTweetText("Hello, World1!");
@@ -39,9 +41,12 @@ public class TweetResource {
 	public void createTweet(Tweet tweet) {
 		String timeStamp = "" + System.currentTimeMillis();
 		tweet.setCreatedAt(timeStamp);
-		
+
 		try {
+			// Save Tweet in DB
 			tweetsDao.createTweet(tweet);
+			// Give it to the Tweet Distribution engine to do the distribution to the followers
+			 
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
